Continue lib/readline work and switch to size_t.
Fixes to lib/readline to update the callback API that it provides
to bash (and other users). Changed from "unsigned int" to size_t
for clarity and to avoid any overhead casting between 32 and 64 bits.
This also fixes build failures on stricter compilers that don't
implicitly cast unsigned int to size_t, e.g. in string append ().

The other major change was to use std::string instead of char *, except
when it required major rewrites that I didn't want to tackle right now.

I also changed "sighandler_t" to "SigHandler" in a few cases, because
Solaris and other systems don't have a typedef for sighandler_t.

+++ lib/readline/undo.cc
@@ -44,12 +44,12 @@ int
 Readline::rl_do_undo ()
 {
   int waiting_for_begin;
-  unsigned int start, end;
+  size_t start, end;
 
 #define TRANS(i)                                                              \
-  ((i) == static_cast<unsigned int> (-1)                                      \
+  ((i) == static_cast<size_t> (-1)                                            \
        ? rl_point                                                             \
-       : ((i) == static_cast<unsigned int> (-2) ? rl_end () : (i)))
+       : ((i) == static_cast<size_t> (-2) ? rl_end () : (i)))
 
   start = end = waiting_for_begin = 0;
   do
@@ -114,7 +114,7 @@ Readline::rl_do_undo ()
 
 /* Save an undo entry for the text from START to END. */
 void
-Readline::rl_modifying (unsigned int start, unsigned int end)
+Readline::rl_modifying (size_t start, size_t end)
 {
   if (start > end)
     {
